:sparkles: Implement routing
